t5s3-epaper: Move variant.cpp -> extra_variants/variant.cpp ...again
Fixes issues with #includes inherited from `configuration.h` when building for pioarduino.

Aligns t5s3_epaper with other variants like t_deck_pro.
